% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_cross_val_fun.R
\name{dog_cross_validation}
\alias{dog_cross_validation}
\title{Cross validation function for dogmatism analysis}
\usage{
dog_cross_validation(data = da, folds_col, model_formula, perc = 0.5,
  above_perc, below_perc, positive)
}
\arguments{
\item{data}{A Twitter data frame}

\item{folds_col}{A fold of unique IDs used to create folds}

\item{model_formula}{the model to test}

\item{perc}{classification percentage, i.e. threshold for when a prediction will be assigned to a specific group}

\item{above_perc}{name assigned when p(x) > perc}

\item{below_perc}{name assigned when p(x) < perc}

\item{positive}{Which factor level is the target.}
}
\value{
returns a 10 by 5 data frame. The five columsn are accuracy, sensitivity,  specificity, auc and fixed effects of each fold.
}
\description{
This function is designed to test discriminability between two groups based on lingustic features related to dogmatic thinking.
The function is desgined to do its work in the context of the dogmatism analytics shiny app. It will need tweaking to work in different
contexts.
The analyis is a 10 fold cross-validation using a binomial logistic regression. Model performence is measured through mean
area-under-the-curve (AUC) of each fold.
}
\examples{

}
