% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_cross_val_fun.R
\name{dog_list_return}
\alias{dog_list_return}
\title{Dogmatism cross validation return}
\usage{
dog_list_return(da, vars, outcome_var = "hashtag")
}
\arguments{
\item{da}{takes a data frame as it's main argument. This function is designed to worked with Twitter data frame used
in the dogmatism analytics Shiny app.}

\item{vars}{a list of variables that will form the predictors in the cross-validation. For dogmatism analysis this
will be close-minded and open-minded words}

\item{outcome_var}{the name of column storing the outcome variable. Mostly likely "hashtag}
}
\description{
dog_list_return returns the results of the dogmatism cross validation in a ggplot2 friendly format.
}
