% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogmrtism.R
\name{dogmrtism}
\alias{dogmrtism}
\title{Analysis of Dogmatism Quotient}
\usage{
dogmrtism(df, col, language = "eng")
}
\arguments{
\item{df}{a dataframe}

\item{col}{the name of text column to be analysis, must be parsed as a string}

\item{language}{only english version is implemented, but a german version exists and will be added later}
}
\value{
returns the original df with three addional columns: open-minded, close-minded, n-words.
}
\description{
This functiosn runs the core dogmatism analysis of text data. It is based on the Dogmatism Quotient Dictionary developed by Suitbert Ertel.
This dictionary contains a list of words related to open-mindedness and close-minded less. The analysis consists of calculating the
proportion of these words
}
\examples{
dog_df <- dogmrtism(test_df, "txt")

}
