% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_extract.R
\name{predictors_extract}
\alias{predictors_extract}
\title{helper function for dogmrtism}
\usage{
predictors_extract(df)
}
\arguments{
\item{df}{}
}
\description{
Extracts close-minded, open-minded and total words for a txt file
}
